+++ ReminderXamarin/ReminderXamarin/ReminderXamarin/ViewModels/AchievementNoteViewModel.cs
@@ -6,7 +6,7 @@ namespace ReminderXamarin.ViewModels
     {
         public int Id { get; set; }
         public string Description { get; set; }
-        public double HoursSpent { get; set; }
+        public int HoursSpent { get; set; }
         public DateTime Date { get; set; }
         public int AchievementId { get; set; }
     }
